Move `previous_proceedings` step to Application task
A bit of code reorganisation as we need to introduce a details step
follow-up to the `previous_proceedings` question and this is not
was part of Abuse concerns task.
