+++ app/services/c100_app/children_decision_tree.rb
@@ -49,7 +49,7 @@ module C100App
       if child_id
         edit(:residence, id: child_id)
       else
-        edit('/steps/abuse_concerns/previous_proceedings')
+        edit('/steps/application/previous_proceedings')
       end
     end
