+++ config/locales/en.yml
@@ -353,11 +353,6 @@ en:
           heading: Contact between your children and the other person
           lead_text: The court needs to know if you will allow your children to have contact with the other person while you wait for a court date.
           info_notice: If you have no safety concerns it’s better for your children to have regular contact with both parents.
-      previous_proceedings:
-        edit:
-          page_title: Previous proceedings
-          heading: Have any of the children in this application been involved in another family court case?
-          lead_text: For example an emergency protection order or supervision order. Include any ongoing cases.
     court_orders:
       has_orders:
         edit:
@@ -552,6 +547,11 @@ en:
           section: Safety concerns
           heading: Provide details of the children’s passports
     application:
+      previous_proceedings:
+        edit:
+          page_title: Previous proceedings
+          heading: Have any of the children in this application been involved in another family court case?
+          lead_text: For example an emergency protection order or supervision order. Include any ongoing cases.
       without_notice:
         edit:
           page_title: Without notice hearing
@@ -750,6 +750,8 @@ en:
     fieldset:
       steps_shared_relationship_form:
         relation_html: ""
+      steps_application_previous_proceedings_form:
+        children_previous_proceedings_html: ""
       steps_application_without_notice_form:
         without_notice_html: ""
       steps_application_without_notice_details_form:
@@ -816,8 +818,6 @@ en:
       steps_abuse_concerns_contact_form:
         concerns_contact_type_html: "Do you agree to your children spending time with the other person?"
         concerns_contact_other_html: "Do you agree to your children having other forms of contact with the other person?"
-      steps_abuse_concerns_previous_proceedings_form:
-        children_previous_proceedings_html: ""
       steps_court_orders_has_orders_form:
         has_court_orders_html: ""
       steps_court_orders_details_form:
