+++ spec/controllers/steps/application/previous_proceedings_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Application::PreviousProceedingsController, type: :controller do
+  it_behaves_like 'an intermediate step controller', Steps::Application::PreviousProceedingsForm, C100App::ApplicationDecisionTree
+end
