+++ spec/forms/steps/application/previous_proceedings_form_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe Steps::AbuseConcerns::PreviousProceedingsForm do
+RSpec.describe Steps::Application::PreviousProceedingsForm do
   it_behaves_like 'a yes-no question form',
                   attribute_name: :children_previous_proceedings
 end
